added a trigger and listener that shows rent movie form with customer list dropdown
