+++ src/views/movie_list_view.js
@@ -4,6 +4,7 @@ import $ from 'jquery';
 import MovieView from './movie_view.js';
 import Movie from '../models/movie.js';
 import MovieList from '../collections/movie_list.js';
+import CustomerListView from './customer_list_view.js';
 
 var MovieListView = Backbone.View.extend ({
   initialize: function(params) {
@@ -25,6 +26,7 @@ var MovieListView = Backbone.View.extend ({
         customers: that.customers
       });
       that.$('#movie-list').append(movieView.render().$el);
+      that.listenTo(movieView, "rentalForm", that.showRentForm);
     });
     return this;
   },
@@ -51,6 +53,20 @@ var MovieListView = Backbone.View.extend ({
       data: {query: [this.getFormData()]}
     });
 
+  },
+  showRentForm: function(movie) {
+    var rentalTemplate = _.template($("#rental-template").html());
+    var compiledRentalTemplate = rentalTemplate(movie.toJSON());
+    $("#create-rental").html(compiledRentalTemplate);
+
+    var rentalCustomerView = new CustomerListView({
+      model: this.customers,
+      template: _.template($('#customer-list-template').html()),
+      tagName: "select",
+      className: "customer-dropdown"
+    });
+
+    $("#create-rental label.customers-select").append(rentalCustomerView.render().$el);
   }
